# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cocolu Documentation",
        "description": "APP Portfolio Management",
        "contact": {
            "email": "daruiza@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080/api",
            "description": "Cocolu API Server"
        }
    ],
    "paths": {
        "/table/{id}/serviceopen": {
            "get": {
                "tags": [
                    "Table"
                ],
                "summary": "Evalua si una tabla tiene Servico y lo retorna",
                "description": "Evalua si una tabla tiene Servico y lo retorna",
                "operationId": "serviceopen",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Table id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get User Token",
                "description": "Return Token",
                "operationId": "getToken",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "GetOut",
                "description": "Close session",
                "operationId": "App\\Http\\Controllers\\Auth\\AuthController::logout",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User body data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the user",
                        "type": "string",
                        "example": "super"
                    },
                    "email": {
                        "title": "email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "super@yopmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password of the user",
                        "type": "string",
                        "example": "0000"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "API EndPoints of Auth"
        },
        {
            "name": "Table",
            "description": "API EndPoints of Table"
        }
    ]
}